requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
pydantic==2.6.4
pydantic-settings==2.2.1
sqlalchemy==2.0.27
psycopg[binary]==3.1.18
pgvector==0.2.4
alembic==1.13.1
torch==2.2.1
torchvision==0.17.1
Pillow==10.2.0
numpy==1.26.4
minio==7.2.7
python-multipart==0.0.9
prometheus-client==0.19.0
httpx==0.27.0
loguru==0.7.2
